requirements.txt
requests
python-dotenv
yfinance
fredapi
pandas
pyarrow
dbt-core
dbt-duckdb
matplotlib
streamlit
apache-airflow
kafka-python
